テストケースを更新

+++ problems/1word-mod/runtime/1word-31/algos/barrett_ceil.hpp
@@ -4,7 +4,7 @@ struct DIV {
  u32 b;
  constexpr DIV(u32 b): b(b), x(u32(-1) / b + 1) {}
  constexpr u32 mod(u32 a) const {
-  u32 r= a - u64(u32(u64(a) * x >> 32)) * b;
+  u32 r= a - u32(u64(a) * x >> 32) * b;
   return r >> 31 ? r + b : r;
  }
 private:
